[IMP] xlsx: import images
This commit implements the functionality of importing images from external xlsx
files. The basic idea is the same as importing charts. The process of unzipping
images and uploading them to the database to get the relative path is
done in Python backend.

This commit also removes `size` in `Image` and move the responsibility
of getting image size to `ImageProvider`.

task 3125702