[IMP] highlight: support unbounded zone
Previously when moving or resizing highlights, the unbounded property of
the zone will not be kept. This commit uses the unbounded zone as the
parameter of the changing highlight command so this property is kept.
Notice that when a zone's bottom is one less than
the sheet height, it will be assumed as the full column. Same thing
applies to the full row.

task 3122611