[FIX] repeat: faster can redo check
With more and more pending revision, `canRedo` was taking more and more
time. Even if it's not that slow (~10ms for 3700 revisions), it's done
at every render! Scrolling feels laggy.

`getRevertedExecution` is not used anymore, but it will be used very
very soon with the version history. So we keep it.