[IMP] spreadsheet: Handle refocus at spreadsheet level
Due to our html structure, we have to support the ability to click on
divs which brings its set of issues concerning the currently focused
element. By default, we want the grid to capture the focus back, and
notably when we click on a top bar/bottom bar button.

To solve this partially, we introduced an effect that would refocus the
grid as it is the default target of all (specifically keyboard) user
interactions. However, the previous implementation could steal the focus
of a component of the library in favor of the focus without the user's
will or intent.
For instance, if user `A` were to edit a side panel input and another
user `B` altered the spreadsheet such that the active sheet of user `A`
would change, they would lose the focus on their input.

This revision moves up the effect to the root component in order to
offer more control on the conditions to force the focus grid;
Specifically, it now has to follow two rules:
1. the current document active element cannot be a child element of the
   Spreadsheet: thanks to this, if a user were to edit an input in the
   side panel, they will no longer see their focus switched without their
   knowledge.

2. Spreadsheet needs to be a child element of the document active
   element: The example provided above could also occur when a user
   focus was on an input outside of the spreadsheet component. This
   condition protects the user as long as the 'focusable' elements
   outside of the spreadsheet are 'parallel' to the latter from a DOM
   tree perspective.

Co-authored-by: Adrien Minne <adrm@odoo.com>